Script to convert HTML in PDF files
During my routine, my supervisor presented a problem. We have an ERP that issues an important document in HTML, this document is sent to the company's customers and they always need to view it correctly.

After a Google update, it is no longer possible to open HTML via Gmail in a clear and visual way, making it necessary to download the file.

This script will be integrated into our ERP, where the conversion from HTML (or HTM) will be done to PDF. This PDF will be sent to customers.